requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.2.0=pypi_0
aioice=0.7.6=pypi_0
aiortc=1.3.2=pypi_0
altair=4.2.0=pypi_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.7.0=pypi_0
av=9.2.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.5=pypi_0
boto3=1.24.91=pypi_0
botocore=1.27.91=pypi_0
bottleneck=1.3.5=py39h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2022.07.19=haa95532_0
cachetools=5.2.0=pypi_0
catalogue=2.0.7=py39haa95532_0
certifi=2022.9.24=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39haa95532_0
colorama=0.4.4=pyhd3eb1b0_0
commonmark=0.9.1=pypi_0
cryptography=37.0.1=py39h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cycler=0.11.0=pyhd3eb1b0_0
cymem=2.0.6=py39hd77b12b_0
cython-blis=0.7.7=py39h080aedc_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
difpy=2.4.4=pypi_0
distlib=0.3.6=pypi_0
dnspython=2.2.1=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.4=py39haa95532_0
et-xmlfile=1.1.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
fastai=2.7.9=py_0
fastcore=1.5.26=py_0
fastdownload=0.0.7=py_0
fastprogress=1.0.3=py_0
fftw=3.3.9=h2bbff1b_1
filelock=3.8.0=pypi_0
fire=0.4.0=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
glib=2.69.1=h5dc1a3c_1
google-auth=2.12.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-crc32c=1.5.0=pypi_0
greenlet=1.1.3=pypi_0
grpcio=1.49.1=pypi_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imageio=2.22.4=pypi_0
importlib-metadata=4.12.0=pypi_0
imutils=0.5.4=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.9.1=py39haa95532_0
ipython=8.3.0=py39haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
jmespath=1.0.1=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
jsonschema=4.4.0=py39haa95532_0
jupyter_client=7.2.2=py39haa95532_0
jupyter_core=4.10.0=py39haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py39hd77b12b_0
langcodes=3.3.0=pyhd3eb1b0_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_4
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libtiff=4.4.0=h8a3f274_0
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.2=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib=3.5.2=py39haa95532_0
matplotlib-base=3.5.2=py39hd77b12b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mediapipe=0.8.11=pypi_0
mistune=0.8.4=py39h2bbff1b_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
msgpack=1.0.4=pypi_0
munkres=1.1.4=py_0
murmurhash=1.0.7=py39hd77b12b_0
nb_conda_kernels=2.3.1=py39haa95532_0
nbclient=0.5.13=py39haa95532_0
nbconvert=6.4.4=py39haa95532_0
nbformat=5.3.0=py39haa95532_0
nest-asyncio=1.5.5=py39haa95532_0
netifaces=0.11.0=pypi_0
networkx=2.8.8=pypi_0
notebook=6.4.12=py39haa95532_0
numexpr=2.8.3=py39hb80d3ca_0
numpy=1.23.5=py39h7a0a035_3
numpy-base=1.21.5=py39hca35cd5_3
oauthlib=3.2.1=pypi_0
opencv-contrib-python=4.6.0.66=pypi_0
opencv-python=4.6.0.66=pypi_0
openpyxl=3.0.10=pypi_0
openssl=1.1.1q=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py39hd77b12b_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathy=0.6.1=py39haa95532_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py39hdc2b20a_1
pip=22.1.2=py39haa95532_0
pipenv=2022.11.11=pypi_0
pipreqs=0.4.11=pypi_0
platformdirs=2.5.4=pypi_0
ply=3.11=py39haa95532_0
preshed=3.0.6=py39h6c2663c_0
prometheus_client=0.14.1=py39haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.6=pypi_0
psutil=5.9.2=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=9.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybboxes=0.1.5=pypi_0
pycodestyle=2.9.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.9.0=py39h2bbff1b_1
pydeck=0.8.0b3=pypi_0
pyee=9.0.4=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pylibsrtp=0.7.1=pypi_0
pympler=1.0.1=pypi_0
pynvim=0.4.3=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39haa95532_0
pyqt=5.15.7=py39hd77b12b_0
pyqt5-sip=12.11.0=py39hd77b12b_0
pyrsistent=0.18.0=py39h196d8e1_0
pysocks=1.7.1=py39haa95532_0
python=3.9.13=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39haa95532_0
pytorch=1.12.1=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py39haa95532_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pywavelets=1.4.1=pypi_0
pywin32=302=py39h2bbff1b_2
pywinpty=2.0.2=py39h5da7b33_0
pyyaml=6.0=py39h2bbff1b_1
pyzmq=22.3.0=py39hd77b12b_2
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=py39haa95532_0
requests-oauthlib=1.3.1=pypi_0
rich=12.5.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.0=pypi_0
sahi=0.10.7=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.1.1=py39hd77b12b_0
scipy=1.7.3=py39h7a0a035_2
seaborn=0.12.0=pypi_0
semver=2.13.0=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=63.4.1=py39haa95532_0
shapely=1.8.5.post1=pypi_0
shellingham=1.3.1=pyhd3eb1b0_0
sip=6.6.2=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0.post1=pypi_0
smart_open=5.2.1=py39haa95532_0
smmap=5.0.0=pypi_0
soupsieve=2.3.1=pyhd3eb1b0_0
spacy=3.3.1=py39h0f1f7c2_0
spacy-legacy=3.0.9=py39haa95532_0
spacy-loggers=1.0.1=pyhd3eb1b0_0
sqlite=3.39.2=h2bbff1b_0
srsly=2.4.3=py39hd77b12b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.12.2=pypi_0
streamlit-webrtc=0.43.4=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.0.1=pypi_0
terminado=0.13.1=py39haa95532_0
terminaltables=3.1.10=pypi_0
testpath=0.6.0=py39haa95532_0
thinc=8.0.15=py39h757b272_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2022.10.10=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pypi_0
torchvision=0.13.1=py39_cu113
tornado=5.1.1=pypi_0
tqdm=4.64.0=py39haa95532_0
traitlets=5.1.1=pyhd3eb1b0_0
typer=0.4.1=py39haa95532_0
typing-extensions=4.3.0=py39haa95532_0
typing_extensions=4.3.0=py39haa95532_0
tzdata=2022.2=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.11=py39haa95532_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
virtualenv=20.16.7=pypi_0
virtualenv-clone=0.5.7=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
wasabi=0.9.1=py39haa95532_0
watchdog=2.1.9=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39haa95532_1
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
winpty=0.4.3=4
xlsxwriter=3.0.3=pypi_0
xz=5.2.5=h8cc25b3_1
yaml=0.2.5=he774522_0
yarg=0.1.9=pypi_0
yolov5=6.2.2=pypi_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h8cc25b3_3
zstd=1.5.2=h19a0ad4_0
